venv-requirements.txt
Flask
numpy
av
ruamel.yaml
PyTurboJPEG